Use a stack instead of a queue for exploring the hierarchy
Factor out a Worklist object for which the underlying data structure can
be easily changed.

* Using a queue, the hierarchy will be explored breadth-first as we
  explore sibling nodes before the current node's children.
* Using a stack, the hierarchy will be explored depth-first as we work
  our way down one subtree before exploring any siblings.

As long as we iterate a given node's children completely before
descending, access patterns should be reasonably favourable for
filesystems. The depth-first search appeared to yield slightly better
performance (~3%, about one second in 30) against a filesystem.

Time for 79359 matches:
* Core 3.1: 19.3 seconds
* .NET 4.8: 25.2 seconds
* .NET 4.6: 32.2 seconds
of which ~80% is I/O.